(>) float ptr_deepest depth b/c of arg.1/2/3...

+++ src/parser.h
@@ -132,7 +132,7 @@ typedef struct Stmt {
 
             // deepest symbol that I point to
             Tk* ptr_deepest_var;
-            int ptr_deepest_depth;  // need to track depth in separate b/c of arg.1/2/3
+            float ptr_deepest_depth;  // need to track depth in separate b/c of arg.1/2/3
         } Var;
         struct {            // STMT_USER
             int  isrec;                 // rec
